improve hazelcast serialization
